fix 5 critical bugs: statistic mismatch, no duration cap, warmup gate missing, BLOB storage, std-zero on restart (#25–#29)
- #25: feed float(row.max()) to baseline instead of np.median(row); detector thresholds
  row.max(), so baseline must track the same statistic — mismatch caused 100% false-positive
  rate on pure noise rows
- #26: add _MAX_ACTIVE_ROWS=600 cap in detector; force-close to POST after 60 s in ACTIVE
  to prevent stuck state + unbounded _frames memory leak
- #27: gate detector.feed() behind baseline.is_warmed_up(); prevents startup noise from
  entering the state machine and producing a bogus event the moment warmup expires
- #28: wrap baseline.mean, baseline.std, snr_db in float() in writer.py; numpy scalars
  stored as 4-byte BLOBs in sqlite3 when not explicitly cast
- #29: add _skip_recompute counter in BaselineTracker.load(); suppresses recompute_std()
  for RING_SIZE//100 calls so the saved std survives until the ring is refreshed with
  real data (avoids std→0 collapse and instant re-trigger on every restart)

+++ src/writer.py
@@ -27,7 +27,7 @@ class EventWriter:
 
         duration_ms = (event.signal_end_time - event.start_time).total_seconds() * 1000
         peak_power_db = float(spectrogram.max())
-        snr_db = peak_power_db - baseline.mean
+        snr_db = float(peak_power_db - baseline.mean)
         integrated_power = float(np.sum(np.maximum(0.0, spectrogram - baseline.mean)))
         frequency_centroid_hz = self._compute_centroid(spectrogram)
         bandwidth_hz = self._compute_bandwidth(spectrogram, frequency_centroid_hz)
@@ -51,8 +51,8 @@ class EventWriter:
                 bandwidth_hz,
                 1 if event.suspected_rfi else 0,
                 cluster_id,
-                baseline.mean,
-                baseline.std,
+                float(baseline.mean),
+                float(baseline.std),
                 spectrogram_blob,
                 spectrogram_shape,
                 _FFT_BIN_WIDTH_HZ,
